pyproject.toml
[project]
name = "biovector-eval"
version = "0.1.0"
description = "A framework for optimizing vector databases in biological/biomedical domains"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "ipywidgets>=8.0.0",
]

[project.scripts]
biovector-eval = "biovector_eval.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biovector_eval"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
# Exclude files with known third-party typing issues
exclude = [
    "create_hmdb_vector_database.py",  # Uses qdrant_client (no stubs)
    "hmdb.py",  # Relative imports for external project
    "tests/",  # Tests use mocks that don't match production types
    "notebooks/",  # Jupyter notebooks have different conventions
]
# Ignore missing imports for libraries without type stubs
reportMissingImports = "warning"
reportMissingTypeStubs = false
# Be lenient with third-party library types (faiss has no stubs)
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
# FAISS library doesn't have type stubs - ignore these errors
reportAttributeAccessIssue = false
reportCallIssue = false
reportOptionalMemberAccess = false

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pyright>=1.1.407",
]
